refactor(core): move CrossChainRecoveryTool to utxo/recovery
Old name was generic and outside the `utxo/` directory

Issue: BG-38773
