+++ modules/core/test/v2/unit/coins/utxo/fixtures/bch/recovery/crossChainRecovery-ltc.json
