+++ modules/core/test/v2/unit/coins/utxo/fixtures/btc/recovery/crossChainRecovery-ltc.json
