+++ modules/core/test/v2/unit/coins/utxo/fixtures/ltc/recovery/crossChainRecovery-btc.json
