+++ modules/core/test/v2/unit/coins/utxo/fixtures/tbch/recovery/crossChainRecovery-tbtc.json
