+++ modules/core/test/v2/unit/coins/utxo/fixtures/tbsv/recovery/crossChainRecovery-tbtc.json
