+++ modules/core/test/v2/unit/coins/utxo/fixtures/tbtc/recovery/crossChainRecovery-tbsv.json
